refactor: create LoginWall parent component for ResetForm and LoginForm
